Conflicts:
	Sources/Internal/Core/Core.cpp
	Sources/Internal/FileSystem/DynamicMemoryFile.h
	Sources/Internal/FileSystem/File.h
	Sources/Internal/Sound/FMODSoundSystem.cpp

Former-commit-id: 92766c83528f0f5d4d9347227e2d44736f219917

+++ Sources/Internal/FileSystem/DynamicMemoryFile.cpp
@@ -66,12 +66,12 @@ uint32 DynamicMemoryFile::Write(const void* pointerToData, uint32 dataSize)
 
     if (data.size() < currentPtr + dataSize)
     {
-        data.resize(currentPtr + dataSize);
+        data.resize(static_cast<size_t>(currentPtr + dataSize));
     }
     if (dataSize)
     {
         DVASSERT(nullptr != pointerToData);
-        Memcpy(&(data[currentPtr]), pointerToData, dataSize);
+        Memcpy(&(data[static_cast<size_t>(currentPtr)]), pointerToData, dataSize);
         currentPtr += dataSize;
     }
 
@@ -92,11 +92,11 @@ uint32 DynamicMemoryFile::Read(void* pointerToData, uint32 dataSize)
     if (currentPtr + realReadSize > size)
     {
         isEof = true;
-        realReadSize = size - currentPtr;
+        realReadSize = size - static_cast<uint32>(currentPtr);
     }
     if (0 < realReadSize)
     {
-        Memcpy(pointerToData, &(data[currentPtr]), realReadSize);
+        Memcpy(pointerToData, &(data[static_cast<size_t>(currentPtr)]), realReadSize);
         currentPtr += realReadSize;
 
         return realReadSize;
@@ -105,19 +105,19 @@ uint32 DynamicMemoryFile::Read(void* pointerToData, uint32 dataSize)
     return 0;
 }
 
-uint32 DynamicMemoryFile::GetPos() const
+uint64 DynamicMemoryFile::GetPos() const
 {
     return currentPtr;
 }
 
-uint32 DynamicMemoryFile::GetSize() const
+uint64 DynamicMemoryFile::GetSize() const
 {
     return static_cast<uint32>(data.size());
 }
 
-bool DynamicMemoryFile::Seek(int32 position, eFileSeek seekType)
+bool DynamicMemoryFile::Seek(int64 position, eFileSeek seekType)
 {
-    int32 pos = 0;
+    int64 pos = 0;
     switch (seekType)
     {
     case SEEK_FROM_START:
